backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.118.0
greenlet==3.2.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx>=0.24,<0.26
idna==3.10
jiter==0.11.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==1.51.2
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
supabase==2.3.4
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
watchfiles==1.1.0
websockets>=11,<13
